examples: re-write in .ts
